Fix a teleport closet in nicolae_vaults_little_squares_5 (various)
An issue I noticed back in 5f48996 (but neglected to handle due to the
scope of that commit), which seems to have been noticed quite a few times
recently, including in #4962. I had briefly thought it was linked to
various Lua update issues, but have since re-discovered the exact vault,
and have dealt with it via some guaranteed door placement.