Properly use seeds for objstat and mapstat
Currently these modes load seed 0 and just keep using the RNG
indefinitely across iterations. This is fine in terms of randomness, but
it means these runs can't be recreated exactly without rerunning exactly
the same number of iterations. When diagnosing map placement errors,
it's helpful to be able to recreate a specific seed.

This commit has objstat and mapstat properly initialize the 0 seed into
a random one and picks a new seed each iteration. Also, for any Lua
errors reported to stderr, we also report the current seed. Finally, we
also respect setting a specific seed via the command line. In this case,
we only allow running one iteration. This could be a useful way to
catalog the items/maps in a specific game seed using these modes.