Fix exploding mounted monsters sometimes crashing
When a mounted monster takes damage, sometimes the rider or the mount
dies leaving the other alive. We handle this by changing the type of the
monster to be the type of the part that survived and then killing a new
monster of the type of the part that died at the same position.
Unfortunately this means that if the code for killing a monster clears
the monster grid at the position of the monster it will leave the living
monster not on the monster grid which can lead to crashes. The monster
on death explosion code was clearing the monster grid as this used to
be needed to prevent the explosion from hitting the dead monster but can
be removed now that explosions are fineffs.