Format multi-line egos/artprops a bit more readably
Now, instead of the descriptions of long egos/artprops spilling over into
the column where the names of those artprops are listed, they will attempt
to be justified in a second 'column'. This should make a number of things
a fair bit more readable, in my opinion.

Unfortunately, since the text of an item description is constructed as a
single giant string, the size of the popup in which is it slated to appear
isn't even to know at the point where text justification occurs - and in
the case of webtiles, I must admit that I have no idea whatsoever how
plausible it is to pass information back to C++ about the size of a popup
handled in html. However, console ensures at least 80 characters are
available and both local tiles and webtiles default to at least that. As
far as I can tell, shrinking the window to the point that this stops being
the case has usually rendered other parts of the UI difficult to use, so
for now I am accepting it as 'good enough' to assume 80 characters are
available.