Fix sticky flame and halo not showing on form tiles (pisastrish)
Tiles for forms that melded all equipment (with the exception of storm
form) were handled in a different place that could only show a single
tile rather than a base tile and multiple tiles for each piece of
'equipment'. Unfortunately, this optimization doesn't really work as the
'equipment' tiles also include the tile for being on fire and the halo
tile behind a glowing character.

Fixes #4979