Fix offhand cleave using primary weapon brand (srcrip, tswn)
Cleaves with offhand weapons had a 50% chance to use the brand of your
primary weapon. This was very noticeable with a spectral weapon in your
primary hand and a non-spectral axe in your offhand as sometimes the
non-spectral axe would place a dancing weapon copy of itself leaving you
two dancing weapons.